poc/requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
matplotlib>=3.8.0
